+++ src/DTF_ESP32Update.cpp
@@ -6,113 +6,96 @@
 #include <WiFiClientSecure.h>
 
 #include "DTF_ESP32Update.h"
+#include <stdint.h>
+
+extern const char* isrgrootx1_cert;
+extern const char* cloudflare_cert;
+extern const char* deploy_the_fleet_cert;
+extern const int CERT_BUNDLE_ID;
 
 static char errstr[128];
+constexpr const int NTP_MAX_WAIT_TIME_IN_SECONDS = 600; // 10 minutes
 
-const char* isrgrootx1_cert PROGMEM = \
-"-----BEGIN CERTIFICATE-----\n" \
-"MIIFazCCA1OgAwIBAgIRAIIQz7DSQONZRGPgu2OCiwAwDQYJKoZIhvcNAQELBQAw\n" \
-"TzELMAkGA1UEBhMCVVMxKTAnBgNVBAoTIEludGVybmV0IFNlY3VyaXR5IFJlc2Vh\n" \
-"cmNoIEdyb3VwMRUwEwYDVQQDEwxJU1JHIFJvb3QgWDEwHhcNMTUwNjA0MTEwNDM4\n" \
-"WhcNMzUwNjA0MTEwNDM4WjBPMQswCQYDVQQGEwJVUzEpMCcGA1UEChMgSW50ZXJu\n" \
-"ZXQgU2VjdXJpdHkgUmVzZWFyY2ggR3JvdXAxFTATBgNVBAMTDElTUkcgUm9vdCBY\n" \
-"MTCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIBAK3oJHP0FDfzm54rVygc\n" \
-"h77ct984kIxuPOZXoHj3dcKi/vVqbvYATyjb3miGbESTtrFj/RQSa78f0uoxmyF+\n" \
-"0TM8ukj13Xnfs7j/EvEhmkvBioZxaUpmZmyPfjxwv60pIgbz5MDmgK7iS4+3mX6U\n" \
-"A5/TR5d8mUgjU+g4rk8Kb4Mu0UlXjIB0ttov0DiNewNwIRt18jA8+o+u3dpjq+sW\n" \
-"T8KOEUt+zwvo/7V3LvSye0rgTBIlDHCNAymg4VMk7BPZ7hm/ELNKjD+Jo2FR3qyH\n" \
-"B5T0Y3HsLuJvW5iB4YlcNHlsdu87kGJ55tukmi8mxdAQ4Q7e2RCOFvu396j3x+UC\n" \
-"B5iPNgiV5+I3lg02dZ77DnKxHZu8A/lJBdiB3QW0KtZB6awBdpUKD9jf1b0SHzUv\n" \
-"KBds0pjBqAlkd25HN7rOrFleaJ1/ctaJxQZBKT5ZPt0m9STJEadao0xAH0ahmbWn\n" \
-"OlFuhjuefXKnEgV4We0+UXgVCwOPjdAvBbI+e0ocS3MFEvzG6uBQE3xDk3SzynTn\n" \
-"jh8BCNAw1FtxNrQHusEwMFxIt4I7mKZ9YIqioymCzLq9gwQbooMDQaHWBfEbwrbw\n" \
-"qHyGO0aoSCqI3Haadr8faqU9GY/rOPNk3sgrDQoo//fb4hVC1CLQJ13hef4Y53CI\n" \
-"rU7m2Ys6xt0nUW7/vGT1M0NPAgMBAAGjQjBAMA4GA1UdDwEB/wQEAwIBBjAPBgNV\n" \
-"HRMBAf8EBTADAQH/MB0GA1UdDgQWBBR5tFnme7bl5AFzgAiIyBpY9umbbjANBgkq\n" \
-"hkiG9w0BAQsFAAOCAgEAVR9YqbyyqFDQDLHYGmkgJykIrGF1XIpu+ILlaS/V9lZL\n" \
-"ubhzEFnTIZd+50xx+7LSYK05qAvqFyFWhfFQDlnrzuBZ6brJFe+GnY+EgPbk6ZGQ\n" \
-"3BebYhtF8GaV0nxvwuo77x/Py9auJ/GpsMiu/X1+mvoiBOv/2X/qkSsisRcOj/KK\n" \
-"NFtY2PwByVS5uCbMiogziUwthDyC3+6WVwW6LLv3xLfHTjuCvjHIInNzktHCgKQ5\n" \
-"ORAzI4JMPJ+GslWYHb4phowim57iaztXOoJwTdwJx4nLCgdNbOhdjsnvzqvHu7Ur\n" \
-"TkXWStAmzOVyyghqpZXjFaH3pO3JLF+l+/+sKAIuvtd7u+Nxe5AW0wdeRlN8NwdC\n" \
-"jNPElpzVmbUq4JUagEiuTDkHzsxHpFKVK7q4+63SM1N95R1NbdWhscdCb+ZAJzVc\n" \
-"oyi3B43njTOQ5yOf+1CceWxG1bQVs5ZufpsMljq4Ui0/1lvh+wjChP4kqKOJ2qxq\n" \
-"4RgqsahDYVvTH9w7jXbyLeiNdd8XM2w9U/t7y0Ff/9yi0GE44Za4rF2LN9d11TPA\n" \
-"mRGunUHBcnWEvgJBQl9nJEiU0Zsnvgc/ubhPgXRR4Xq37Z0j4r7g1SgEEzwxA57d\n" \
-"emyPxgcYxn/eR44/KJ4EBs+lVDR3veyJm+kXQ99b21/+jh5Xos1AnX5iItreGCc=\n" \
-"-----END CERTIFICATE-----\n";
-
-const char* cloudflare_cert PROGMEM = \
-"-----BEGIN CERTIFICATE-----\n" \
-"MIIEADCCAuigAwIBAgIID+rOSdTGfGcwDQYJKoZIhvcNAQELBQAwgYsxCzAJBgNV\n" \
-"BAYTAlVTMRkwFwYDVQQKExBDbG91ZEZsYXJlLCBJbmMuMTQwMgYDVQQLEytDbG91\n" \
-"ZEZsYXJlIE9yaWdpbiBTU0wgQ2VydGlmaWNhdGUgQXV0aG9yaXR5MRYwFAYDVQQH\n" \
-"Ew1TYW4gRnJhbmNpc2NvMRMwEQYDVQQIEwpDYWxpZm9ybmlhMB4XDTE5MDgyMzIx\n" \
-"MDgwMFoXDTI5MDgxNTE3MDAwMFowgYsxCzAJBgNVBAYTAlVTMRkwFwYDVQQKExBD\n" \
-"bG91ZEZsYXJlLCBJbmMuMTQwMgYDVQQLEytDbG91ZEZsYXJlIE9yaWdpbiBTU0wg\n" \
-"Q2VydGlmaWNhdGUgQXV0aG9yaXR5MRYwFAYDVQQHEw1TYW4gRnJhbmNpc2NvMRMw\n" \
-"EQYDVQQIEwpDYWxpZm9ybmlhMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKC\n" \
-"AQEAwEiVZ/UoQpHmFsHvk5isBxRehukP8DG9JhFev3WZtG76WoTthvLJFRKFCHXm\n" \
-"V6Z5/66Z4S09mgsUuFwvJzMnE6Ej6yIsYNCb9r9QORa8BdhrkNn6kdTly3mdnykb\n" \
-"OomnwbUfLlExVgNdlP0XoRoeMwbQ4598foiHblO2B/LKuNfJzAMfS7oZe34b+vLB\n" \
-"yrP/1bgCSLdc1AxQc1AC0EsQQhgcyTJNgnG4va1c7ogPlwKyhbDyZ4e59N5lbYPJ\n" \
-"SmXI/cAe3jXj1FBLJZkwnoDKe0v13xeF+nF32smSH0qB7aJX2tBMW4TWtFPmzs5I\n" \
-"lwrFSySWAdwYdgxw180yKU0dvwIDAQABo2YwZDAOBgNVHQ8BAf8EBAMCAQYwEgYD\n" \
-"VR0TAQH/BAgwBgEB/wIBAjAdBgNVHQ4EFgQUJOhTV118NECHqeuU27rhFnj8KaQw\n" \
-"HwYDVR0jBBgwFoAUJOhTV118NECHqeuU27rhFnj8KaQwDQYJKoZIhvcNAQELBQAD\n" \
-"ggEBAHwOf9Ur1l0Ar5vFE6PNrZWrDfQIMyEfdgSKofCdTckbqXNTiXdgbHs+TWoQ\n" \
-"wAB0pfJDAHJDXOTCWRyTeXOseeOi5Btj5CnEuw3P0oXqdqevM1/+uWp0CM35zgZ8\n" \
-"VD4aITxity0djzE6Qnx3Syzz+ZkoBgTnNum7d9A66/V636x4vTeqbZFBr9erJzgz\n" \
-"hhurjcoacvRNhnjtDRM0dPeiCJ50CP3wEYuvUzDHUaowOsnLCjQIkWbR7Ni6KEIk\n" \
-"MOz2U0OBSif3FTkhCgZWQKOOLo1P42jHC3ssUZAtVNXrCk3fw9/E15k8NPkBazZ6\n" \
-"0iykLhH1trywrKRMVw67F44IE8Y=\n" \
-"-----END CERTIFICATE-----\n";
+enum CertificateType {
+    ISRG_ROOT_X1,
+    CLOUDFLARE_ORIGIN,
+    DEPLOY_THE_FLEET_X1
+};
 
 void setSystemTime()
 {
     configTime(0, 0, "pool.ntp.org", "time.nist.gov");
 
-    #ifdef DTF_DEBUG
-    Serial.print(F("Waiting for NTP time sync: "));
-    #endif
+    log_i("Waiting for NTP time sync: ");
 
     time_t now = time(nullptr);
-    while (now < 8 * 3600 * 2) {
+    while (now < NTP_MAX_WAIT_TIME_IN_SECONDS) {
         yield();
         delay(500);
-        #ifdef DTF_DEBUG
-        Serial.print(F("."));
-        #endif
+        log_d(".");
         now = time(nullptr);
     }
 
-    #ifdef DTF_DEBUG
-    Serial.println(F(""));
-    #endif
     struct tm timeinfo;
     gmtime_r(&now, &timeinfo);
 
-    #ifdef DTF_DEBUG
-    Serial.print(F("Current time: "));
-    Serial.print(asctime(&timeinfo));
-    #endif
+    log_i("Current time: %s", asctime(&timeinfo));
 }
 
-DTF_UpdateResponse DTF_ESP32Update::getFirmwareUpdate(const char* updateUrl, const char* currentVersion, bool setTime)
+void setCertificate(WiFiClientSecure& client, CertificateType certType) {
+    switch (certType) {
+        case ISRG_ROOT_X1:
+            log_d("Setting certificate to ISRG Root X1");
+            client.setCACert(isrgrootx1_cert);
+            break;
+        case CLOUDFLARE_ORIGIN:
+            log_d("Setting certificate to Cloudflare Origin");
+            client.setCACert(cloudflare_cert);
+            break;
+        case DEPLOY_THE_FLEET_X1:
+            log_d("Setting certificate to Deploy the Fleet X1");
+            client.setCACert(deploy_the_fleet_cert);
+            break;
+    }
+}
+
+DTF_UpdateResponse DTF_ESP32Update::getFirmwareUpdate(
+    const char* updateUrl, 
+    const char* currentVersion, 
+    DTF_RebootOption rebootOption, 
+    DTF_SetTimeOption setTimeOption
+)
 {
-    #ifdef DTF_DEBUG
-    Serial.println("Checking for firmware updates");
-    #endif
+    log_i("Checking for firmware updates");
+
+    // Append the certificate bundle ID to the update URL
+    String url = String(updateUrl);
+    url.reserve(strlen(updateUrl) + 12);
+    if(url.indexOf("?") == -1)
+    {
+        url.concat("?cb=");
+    }
+    else
+    {
+        url.concat("&cb=");
+    }
+    url.concat(CERT_BUNDLE_ID);
 
     // Ensure the current time is set and accurate
-    if (setTime)
+    if (setTimeOption == DTF_SetTimeOption::SET_TIME)
     {
         setSystemTime();
     }
 
-    // Create a secure client using the Let's Encrypt certificate
+    if(rebootOption == DTF_RebootOption::NO_REBOOT)
+    {
+        httpUpdate.rebootOnUpdate(false);
+    }
+
+    // Create a secure client using the Let's Encrypt X1 root certificate
     WiFiClientSecure client;
-    client.setCACert(isrgrootx1_cert);
+    log_d("Setting certificate to ISRG Root X1");
+    setCertificate(client, CertificateType::ISRG_ROOT_X1);
     client.setTimeout(20); // 20 second timeout
 
     // The following line invokes the update library and will send all necessary headers
@@ -123,49 +106,36 @@ DTF_UpdateResponse DTF_ESP32Update::getFirmwareUpdate(const char* updateUrl, con
     int lastErr = client.lastError(errstr, sizeof(errstr));
 
     if(ret == HTTP_UPDATE_FAILED){
-        #ifdef DTF_DEBUG
-        Serial.println("Update failed.");
-        Serial.print("Last error: ");
-        Serial.print(lastErr);
-        Serial.println(errstr);
-        Serial.println("Trying again with backup certificate");
-        #endif
-
-        // Try again with a backup certificate
-        client.setCACert(cloudflare_cert);
+        log_d("Update failed. Last error: %d %s", lastErr, errstr);
+        setCertificate(client, CertificateType::CLOUDFLARE_ORIGIN);
         ret = httpUpdate.update(client, updateUrl, currentVersion);
 
         if(ret == HTTP_UPDATE_FAILED){
             lastErr = client.lastError(errstr, sizeof(errstr));
-            #ifdef DTF_DEBUG
-            Serial.println("Update failed with backup certificate.");
-            Serial.print("Last error: ");
-            Serial.print(lastErr);
-            Serial.println(errstr);
-            Serial.println("Trying again with backup certificate");
-            #endif
+            log_d("Update failed. Last error: %d %s", lastErr, errstr);
+
+            // Change the domain to the failover
+            url.replace("ota.", "ota9.");
+            log_d("Backup URL: %s", url.c_str());
+            setCertificate(client, CertificateType::DEPLOY_THE_FLEET_X1);
+            ret = httpUpdate.update(client, url.c_str(), currentVersion);
         }
     }
 
     switch (ret) {
         case HTTP_UPDATE_FAILED:
-            #ifdef DTF_DEBUG
-            Serial.printf("HTTP_UPDATE_FAILED Error (%d): %s\n", httpUpdate.getLastError(), httpUpdate.getLastErrorString().c_str());
-            #endif
+            lastErr = client.lastError(errstr, sizeof(errstr));
+            log_e("Update failed. Last error: %d %s", lastErr, errstr);
             return DTF_UpdateResponse::UPDATE_FAILED;
             break;
 
         case HTTP_UPDATE_NO_UPDATES:
-            #ifdef DTF_DEBUG
-            Serial.println("No updates are available for this device.");
-            #endif
+            log_i("No updates are available for this device.");
             return DTF_UpdateResponse::NO_UPDATE_AVAILABLE;
             break;
 
         case HTTP_UPDATE_OK:
-            #ifdef DTF_DEBUG
-            Serial.println("Update succeeded.");
-            #endif
+            log_i("Update succeeded. Reboot device to apply update.");
             return DTF_UpdateResponse::UPDATE_SUCCEEDED;
             break;